refactor(web): extract color themes to JSON and consolidate context files
Move 24 color theme definitions from TypeScript (colorThemes.ts) into a
static JSON data file (themes.json), reducing colorThemes.ts from 286 to
31 lines. Consolidate ThemeContextDef.ts, themeStorage.ts, fontLoader.ts,
and localeStorage.ts into ThemeContext.tsx to reduce file count and
TypeScript footprint. Net: -282 TypeScript lines.
