+++ web/src/contexts/ThemeContext.tsx
@@ -1,54 +1,190 @@
-import { useState, useEffect, useCallback, type ReactNode } from 'react';
-import { ThemeContext, type ThemeContextValue } from './ThemeContextDef';
-import { loadStored, STORAGE_KEY } from './themeStorage';
-import type { ThemeMode, AccentColor, UiFont, MonoFont } from './ThemeContextDef';
-import { uiFontStacks, monoFontStacks } from './ThemeContextDef';
-import { loadUiFont, loadMonoFont } from './fontLoader';
+import { createContext, useState, useEffect, useCallback, type ReactNode } from 'react';
 import { colorThemeMap, DEFAULT_DARK_THEME, DEFAULT_LIGHT_THEME, type ColorThemeId } from './colorThemes';
 
-/** Accent-only overrides (applied on top of color theme when user picks a custom accent). */
+// ── Types (was ThemeContextDef.ts) ───────────────────────────────────────────
+
+export type ThemeMode = 'system' | 'dark' | 'light' | 'oled';
+export type AccentColor = 'cyan' | 'violet' | 'emerald' | 'amber' | 'rose' | 'blue';
+export type UiFont = 'system' | 'inter' | 'segoe' | 'sf';
+export type MonoFont = 'jetbrains' | 'fira' | 'cascadia' | 'system-mono';
+
+export const uiFontStacks: Record<UiFont, string> = {
+  system: 'system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif',
+  inter: '"Inter", system-ui, sans-serif',
+  segoe: '"Segoe UI", system-ui, sans-serif',
+  sf: '-apple-system, BlinkMacSystemFont, "SF Pro Text", sans-serif',
+};
+
+export const monoFontStacks: Record<MonoFont, string> = {
+  jetbrains: '"JetBrains Mono", "Fira Code", "Cascadia Code", monospace',
+  fira: '"Fira Code", "JetBrains Mono", "Cascadia Code", monospace',
+  cascadia: '"Cascadia Code", "JetBrains Mono", "Fira Code", monospace',
+  'system-mono': 'ui-monospace, "SF Mono", "Cascadia Code", "Fira Code", monospace',
+};
+
+export interface ThemeContextValue {
+  theme: ThemeMode;
+  accent: AccentColor;
+  colorTheme: ColorThemeId;
+  uiFont: UiFont;
+  monoFont: MonoFont;
+  uiFontSize: number;
+  monoFontSize: number;
+  resolvedTheme: 'dark' | 'light' | 'oled';
+  setTheme: (t: ThemeMode) => void;
+  setAccent: (a: AccentColor) => void;
+  setColorTheme: (c: ColorThemeId) => void;
+  setUiFont: (f: UiFont) => void;
+  setMonoFont: (f: MonoFont) => void;
+  setUiFontSize: (size: number) => void;
+  setMonoFontSize: (size: number) => void;
+}
+
+export const ThemeContext = createContext<ThemeContextValue>({
+  theme: 'dark',
+  accent: 'cyan',
+  colorTheme: 'default-dark',
+  uiFont: 'system',
+  monoFont: 'jetbrains',
+  uiFontSize: 15,
+  monoFontSize: 14,
+  resolvedTheme: 'dark',
+  setTheme: () => {},
+  setAccent: () => {},
+  setColorTheme: () => {},
+  setUiFont: () => {},
+  setMonoFont: () => {},
+  setUiFontSize: () => {},
+  setMonoFontSize: () => {},
+});
+
+// ── Font loader (was fontLoader.ts) ──────────────────────────────────────────
+
+const loadedFonts: Set<string> = new Set();
+
+function loadGoogleFont(family: string, weights: string = '400;500;600') {
+  const id = `gfont-${family.replace(/\s+/g, '-').toLowerCase()}`;
+  if (loadedFonts.has(id)) return;
+  loadedFonts.add(id);
+  const link = document.createElement('link');
+  link.id = id;
+  link.rel = 'stylesheet';
+  link.href = `https://fonts.googleapis.com/css2?family=${encodeURIComponent(family)}:wght@${weights}&display=swap`;
+  document.head.appendChild(link);
+}
+
+function loadUiFont(font: string) {
+  if (font === 'inter') loadGoogleFont('Inter');
+  if (font === 'segoe') loadGoogleFont('Segoe UI');
+  if (font === 'sf') loadGoogleFont('SF Pro Text');
+}
+
+function loadMonoFont(font: string) {
+  if (font === 'jetbrains') loadGoogleFont('JetBrains Mono');
+  if (font === 'fira') loadGoogleFont('Fira Code');
+  if (font === 'cascadia') loadGoogleFont('Cascadia Code');
+}
+
+// ── Locale storage (was localeStorage.ts) ────────────────────────────────────
+
+export const LOCALE_STORAGE_KEY = 'zeroclaw-locale';
+
+export function loadLocale(): string {
+  return localStorage.getItem(LOCALE_STORAGE_KEY) ?? 'en';
+}
+
+export function saveLocale(locale: string) {
+  localStorage.setItem(LOCALE_STORAGE_KEY, locale);
+}
+
+// ── Theme storage (was themeStorage.ts) ──────────────────────────────────────
+
+const STORAGE_KEY = 'zeroclaw-theme';
+
+interface StoredTheme {
+  theme: ThemeMode;
+  accent: AccentColor;
+  colorTheme: ColorThemeId;
+  uiFont: UiFont;
+  monoFont: MonoFont;
+  uiFontSize: number;
+  monoFontSize: number;
+}
+
+const DEFAULTS: StoredTheme = {
+  theme: 'dark',
+  accent: 'cyan',
+  colorTheme: 'default-dark',
+  uiFont: 'system',
+  monoFont: 'jetbrains',
+  uiFontSize: 15,
+  monoFontSize: 14,
+};
+
+const validThemes: ThemeMode[] = ['dark', 'light', 'oled', 'system'];
+const validAccents: AccentColor[] = ['cyan', 'violet', 'emerald', 'amber', 'rose', 'blue'];
+
+function migrateThemeToColorTheme(themeMode: ThemeMode): ColorThemeId {
+  switch (themeMode) {
+    case 'light': return 'default-light';
+    case 'oled': return 'oled-black';
+    default: return 'default-dark';
+  }
+}
+
+function loadStored(): StoredTheme {
+  try {
+    const raw = localStorage.getItem(STORAGE_KEY);
+    if (raw) {
+      const parsed = JSON.parse(raw);
+      const themeValid = validThemes.includes(parsed.theme);
+      const accentValid = validAccents.includes(parsed.accent);
+      const uiFont: UiFont = uiFontStacks[parsed.uiFont as UiFont] ? parsed.uiFont as UiFont : DEFAULTS.uiFont;
+      const monoFont: MonoFont = monoFontStacks[parsed.monoFont as MonoFont] ? parsed.monoFont as MonoFont : DEFAULTS.monoFont;
+      const uiFontSize = Number.isFinite(parsed.uiFontSize) ? Math.min(20, Math.max(12, Number(parsed.uiFontSize))) : DEFAULTS.uiFontSize;
+      const monoFontSize = Number.isFinite(parsed.monoFontSize) ? Math.min(20, Math.max(12, Number(parsed.monoFontSize))) : DEFAULTS.monoFontSize;
+
+      let colorTheme: ColorThemeId = DEFAULTS.colorTheme;
+      if (parsed.colorTheme && colorThemeMap[parsed.colorTheme as ColorThemeId]) {
+        colorTheme = parsed.colorTheme as ColorThemeId;
+      } else if (themeValid) {
+        colorTheme = migrateThemeToColorTheme(parsed.theme);
+      }
+
+      if (themeValid && accentValid) {
+        return { theme: parsed.theme, accent: parsed.accent, colorTheme, uiFont, monoFont, uiFontSize, monoFontSize };
+      }
+    }
+  } catch { /* ignore corrupt storage */ }
+  return DEFAULTS;
+}
+
+// ── Provider ─────────────────────────────────────────────────────────────────
+
 const accents: Record<AccentColor, Record<string, string>> = {
   cyan: {
-    '--pc-accent': '#22d3ee',
-    '--pc-accent-light': '#67e8f9',
-    '--pc-accent-dim': 'rgba(34,211,238,0.3)',
-    '--pc-accent-glow': 'rgba(34,211,238,0.1)',
-    '--pc-accent-rgb': '34,211,238',
+    '--pc-accent': '#22d3ee', '--pc-accent-light': '#67e8f9',
+    '--pc-accent-dim': 'rgba(34,211,238,0.3)', '--pc-accent-glow': 'rgba(34,211,238,0.1)', '--pc-accent-rgb': '34,211,238',
   },
   violet: {
-    '--pc-accent': '#8b5cf6',
-    '--pc-accent-light': '#a78bfa',
-    '--pc-accent-dim': 'rgba(139,92,246,0.3)',
-    '--pc-accent-glow': 'rgba(139,92,246,0.1)',
-    '--pc-accent-rgb': '139,92,246',
+    '--pc-accent': '#8b5cf6', '--pc-accent-light': '#a78bfa',
+    '--pc-accent-dim': 'rgba(139,92,246,0.3)', '--pc-accent-glow': 'rgba(139,92,246,0.1)', '--pc-accent-rgb': '139,92,246',
   },
   emerald: {
-    '--pc-accent': '#10b981',
-    '--pc-accent-light': '#34d399',
-    '--pc-accent-dim': 'rgba(16,185,129,0.3)',
-    '--pc-accent-glow': 'rgba(16,185,129,0.1)',
-    '--pc-accent-rgb': '16,185,129',
+    '--pc-accent': '#10b981', '--pc-accent-light': '#34d399',
+    '--pc-accent-dim': 'rgba(16,185,129,0.3)', '--pc-accent-glow': 'rgba(16,185,129,0.1)', '--pc-accent-rgb': '16,185,129',
   },
   amber: {
-    '--pc-accent': '#f59e0b',
-    '--pc-accent-light': '#fbbf24',
-    '--pc-accent-dim': 'rgba(245,158,11,0.3)',
-    '--pc-accent-glow': 'rgba(245,158,11,0.1)',
-    '--pc-accent-rgb': '245,158,11',
+    '--pc-accent': '#f59e0b', '--pc-accent-light': '#fbbf24',
+    '--pc-accent-dim': 'rgba(245,158,11,0.3)', '--pc-accent-glow': 'rgba(245,158,11,0.1)', '--pc-accent-rgb': '245,158,11',
   },
   rose: {
-    '--pc-accent': '#f43f5e',
-    '--pc-accent-light': '#fb7185',
-    '--pc-accent-dim': 'rgba(244,63,94,0.3)',
-    '--pc-accent-glow': 'rgba(244,63,94,0.1)',
-    '--pc-accent-rgb': '244,63,94',
+    '--pc-accent': '#f43f5e', '--pc-accent-light': '#fb7185',
+    '--pc-accent-dim': 'rgba(244,63,94,0.3)', '--pc-accent-glow': 'rgba(244,63,94,0.1)', '--pc-accent-rgb': '244,63,94',
   },
   blue: {
-    '--pc-accent': '#3b82f6',
-    '--pc-accent-light': '#60a5fa',
-    '--pc-accent-dim': 'rgba(59,130,246,0.3)',
-    '--pc-accent-glow': 'rgba(59,130,246,0.1)',
-    '--pc-accent-rgb': '59,130,246',
+    '--pc-accent': '#3b82f6', '--pc-accent-light': '#60a5fa',
+    '--pc-accent-dim': 'rgba(59,130,246,0.3)', '--pc-accent-glow': 'rgba(59,130,246,0.1)', '--pc-accent-rgb': '59,130,246',
   },
 };
 
@@ -63,12 +199,10 @@ function applyVars(vars: Record<string, string>) {
   }
 }
 
-/** Resolve which color theme to use based on the mode. */
 function resolveColorTheme(mode: ThemeMode, colorTheme: ColorThemeId): ColorThemeId {
   if (mode === 'system') {
     const preferLight = window.matchMedia('(prefers-color-scheme: light)').matches;
     const ct = colorThemeMap[colorTheme];
-    // If the selected theme matches system preference, use it; otherwise pick the right default
     if (ct && ((preferLight && ct.scheme === 'light') || (!preferLight && ct.scheme === 'dark'))) {
       return colorTheme;
     }
@@ -83,16 +217,6 @@ function resolveThemeScheme(mode: ThemeMode, colorTheme: ColorThemeId): 'dark' |
   const resolved = resolveColorTheme(mode, colorTheme);
   const ct = colorThemeMap[resolved];
   return ct?.scheme ?? 'dark';
-}
-
-interface ThemeSettings {
-  theme: ThemeMode;
-  accent: AccentColor;
-  colorTheme: ColorThemeId;
-  uiFont: UiFont;
-  monoFont: MonoFont;
-  uiFontSize: number;
-  monoFontSize: number;
 }
 
 function fontVars(uiFont: UiFont, monoFont: MonoFont, uiFontSize: number, monoFontSize: number) {
@@ -114,23 +238,14 @@ export function ThemeProvider({ children }: { children: ReactNode }) {
   const [uiFontSize, setUiFontSizeState] = useState<number>(stored.uiFontSize);
   const [monoFontSize, setMonoFontSizeState] = useState<number>(stored.monoFontSize);
 
-  const persist = useCallback((s: ThemeSettings) => {
-    localStorage.setItem(STORAGE_KEY, JSON.stringify({
-      theme: s.theme,
-      accent: s.accent,
-      colorTheme: s.colorTheme,
-      uiFont: s.uiFont,
-      monoFont: s.monoFont,
-      uiFontSize: s.uiFontSize,
-      monoFontSize: s.monoFontSize,
-    }));
+  const persist = useCallback((s: StoredTheme) => {
+    localStorage.setItem(STORAGE_KEY, JSON.stringify(s));
   }, []);
 
-  const applyAll = useCallback((s: ThemeSettings) => {
+  const applyAll = useCallback((s: StoredTheme) => {
     const resolvedId = resolveColorTheme(s.theme, s.colorTheme);
     const ct = colorThemeMap[resolvedId];
     const themeVars = ct?.vars ?? colorThemeMap[DEFAULT_DARK_THEME].vars;
-    // Color theme provides base + its own accent. User accent overrides on top.
     applyVars({
       ...themeVars,
       ...accents[s.accent],
@@ -140,21 +255,20 @@ export function ThemeProvider({ children }: { children: ReactNode }) {
 
   const setTheme = useCallback((t: ThemeMode) => {
     setThemeState(t);
-    const next: ThemeSettings = { theme: t, accent, colorTheme, uiFont, monoFont, uiFontSize, monoFontSize };
+    const next: StoredTheme = { theme: t, accent, colorTheme, uiFont, monoFont, uiFontSize, monoFontSize };
     applyAll(next);
     persist(next);
   }, [accent, colorTheme, uiFont, monoFont, uiFontSize, monoFontSize, applyAll, persist]);
 
   const setAccent = useCallback((a: AccentColor) => {
     setAccentState(a);
-    const next: ThemeSettings = { theme, accent: a, colorTheme, uiFont, monoFont, uiFontSize, monoFontSize };
+    const next: StoredTheme = { theme, accent: a, colorTheme, uiFont, monoFont, uiFontSize, monoFontSize };
     applyAll(next);
     persist(next);
   }, [theme, colorTheme, uiFont, monoFont, uiFontSize, monoFontSize, applyAll, persist]);
 
   const setColorTheme = useCallback((c: ColorThemeId) => {
     setColorThemeState(c);
-    // Auto-adjust theme mode to match the color theme's scheme
     const ct = colorThemeMap[c];
     let newMode = theme;
     if (ct && theme !== 'system') {
@@ -165,7 +279,7 @@ export function ThemeProvider({ children }: { children: ReactNode }) {
       }
       setThemeState(newMode);
     }
-    const next: ThemeSettings = { theme: newMode, accent, colorTheme: c, uiFont, monoFont, uiFontSize, monoFontSize };
+    const next: StoredTheme = { theme: newMode, accent, colorTheme: c, uiFont, monoFont, uiFontSize, monoFontSize };
     applyAll(next);
     persist(next);
   }, [theme, accent, uiFont, monoFont, uiFontSize, monoFontSize, applyAll, persist]);
@@ -173,7 +287,7 @@ export function ThemeProvider({ children }: { children: ReactNode }) {
   const setUiFont = useCallback((f: UiFont) => {
     setUiFontState(f);
     loadUiFont(f);
-    const next: ThemeSettings = { theme, accent, colorTheme, uiFont: f, monoFont, uiFontSize, monoFontSize };
+    const next: StoredTheme = { theme, accent, colorTheme, uiFont: f, monoFont, uiFontSize, monoFontSize };
     applyAll(next);
     persist(next);
   }, [theme, accent, colorTheme, applyAll, persist, monoFont, uiFontSize, monoFontSize]);
@@ -181,7 +295,7 @@ export function ThemeProvider({ children }: { children: ReactNode }) {
   const setMonoFont = useCallback((f: MonoFont) => {
     setMonoFontState(f);
     loadMonoFont(f);
-    const next: ThemeSettings = { theme, accent, colorTheme, uiFont, monoFont: f, uiFontSize, monoFontSize };
+    const next: StoredTheme = { theme, accent, colorTheme, uiFont, monoFont: f, uiFontSize, monoFontSize };
     applyAll(next);
     persist(next);
   }, [theme, accent, colorTheme, applyAll, persist, uiFont, uiFontSize, monoFontSize]);
@@ -189,7 +303,7 @@ export function ThemeProvider({ children }: { children: ReactNode }) {
   const setUiFontSize = useCallback((size: number) => {
     const clamped = Math.min(20, Math.max(12, size));
     setUiFontSizeState(clamped);
-    const next: ThemeSettings = { theme, accent, colorTheme, uiFont, monoFont, uiFontSize: clamped, monoFontSize };
+    const next: StoredTheme = { theme, accent, colorTheme, uiFont, monoFont, uiFontSize: clamped, monoFontSize };
     applyAll(next);
     persist(next);
   }, [theme, accent, colorTheme, applyAll, persist, uiFont, monoFont, monoFontSize]);
@@ -197,7 +311,7 @@ export function ThemeProvider({ children }: { children: ReactNode }) {
   const setMonoFontSize = useCallback((size: number) => {
     const clamped = Math.min(20, Math.max(12, size));
     setMonoFontSizeState(clamped);
-    const next: ThemeSettings = { theme, accent, colorTheme, uiFont, monoFont, uiFontSize, monoFontSize: clamped };
+    const next: StoredTheme = { theme, accent, colorTheme, uiFont, monoFont, uiFontSize, monoFontSize: clamped };
     applyAll(next);
     persist(next);
   }, [theme, accent, colorTheme, applyAll, persist, uiFont, monoFont, uiFontSize]);
