+++ web/src/hooks/useTheme.ts
@@ -1,4 +1,4 @@
 import { useContext } from 'react';
-import { ThemeContext } from '../contexts/ThemeContextDef';
+import { ThemeContext } from '../contexts/ThemeContext';
 
 export const useTheme = () => useContext(ThemeContext);
